+++ src/lib_macos.c
@@ -279,7 +279,7 @@ int m_accept(int socket, void *clientAddr, socklen_t clientAddrSize)
     return r;
 }
 
-int m_recv(int socket, void *buf, int len)
+ssize_t m_recv(int socket, void *buf, size_t len)
 {
     int r = recv(socket, buf, len, 0);
     if (unlikely(r == -1))
@@ -289,7 +289,7 @@ int m_recv(int socket, void *buf, int len)
     return r;
 }
 
-int m_send(int socket, void *buf, int len)
+ssize_t m_send(int socket, void *buf, size_t len)
 {
     int r = send(socket, buf, len, 0);
     if (unlikely(r == -1))
